second commit

+++ app/src/main/java/com/example/food_wastage_management/MyDatabaseHelper.java
@@ -7,12 +7,9 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.widget.Toast;
 
-import androidx.annotation.Nullable;
-
 class MyDatabaseHelper extends SQLiteOpenHelper {
-    private Context context;
     private static final String DATABASE_NAME = "food-management.db";
-    private static final int DATABASE_VERSION =1;
+    private static final int DATABASE_VERSION = 2; // Increment for version update
     private static final String TABLE_NAME = "my_foods";
     private static final String COLUMN_ID = "customer_id";
     private static final String COLUMN_NAME = "customer_name";
@@ -23,32 +20,24 @@ class MyDatabaseHelper extends SQLiteOpenHelper {
     private static final String COLUMN_FOODTYPE = "food_type";
     private static final String COLUMN_FOODQUANTITY = "food_quantity";
 
-
-    public MyDatabaseHelper(@Nullable Context context) {
+    public MyDatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
-        this.context=context;
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        String query = "CREATE TABLE " + TABLE_NAME + " (" +
-                COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
-                COLUMN_NAME + " TEXT NOT NULL, " +
-                COLUMN_EMAIL + " TEXT NOT NULL, " +
-                COLUMN_PHNO + " TEXT NOT NULL, " +
-                COLUMN_ADDRESS + " TEXT NOT NULL, " +
-                COLUMN_ZIP + " TEXT NOT NULL, " +
-                COLUMN_FOODTYPE + " TEXT, " +
-                COLUMN_FOODQUANTITY + " INTEGER DEFAULT 0);";
-        db.execSQL(query);
+        db.execSQL("CREATE TABLE " + TABLE_NAME + " (" + COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " + COLUMN_NAME + " TEXT NOT NULL, " + COLUMN_EMAIL + " TEXT NOT NULL, " + COLUMN_PHNO + " INTEGER NOT NULL, " + COLUMN_ADDRESS + " TEXT NOT NULL, " + COLUMN_ZIP + " INTEGER NOT NULL, " + COLUMN_FOODTYPE + " TEXT, " + COLUMN_FOODQUANTITY + " INTEGER DEFAULT 0 NOT NULL);");
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase db, int i, int i1) {
-        db.execSQL("DROP TABLE IF EXISTS "+TABLE_NAME);
-        onCreate(db);
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(db); // Recreate the table with updated schema
+        }
     }
-    void addFood(String name, String email, int phno, String address, int zip, String foodtype, int foodquantity) {
+
+    public boolean addFood(String name, String email, String phno, String address, String zip, String foodtype, String foodquantity) {
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues cv = new ContentValues();
 
@@ -59,23 +48,25 @@ class MyDatabaseHelper extends SQLiteOpenHelper {
         cv.put(COLUMN_ZIP, zip);
         cv.put(COLUMN_FOODTYPE, foodtype);
         cv.put(COLUMN_FOODQUANTITY, foodquantity);
-        long result = db.insert(TABLE_NAME, null,cv);
-        if (result == -1) {
-            Toast.makeText(context, "Failed to submit!", Toast.LENGTH_LONG).show();
-        } else {
-            Toast.makeText(context, "Submitted Successfully!", Toast.LENGTH_LONG).show();
-        }
 
+
+        long result = db.insert(TABLE_NAME, null, cv);
+        if (result != -1) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
-    Cursor readAllData() {
-        String query = "SELECT * FROM " + TABLE_NAME;
-        SQLiteDatabase db = this.getReadableDatabase();
-
-        Cursor cursor = null;
-        if (db!=null) {
-            cursor = db.rawQuery(query, null);
-        }
+    public Cursor readAllData() {
+        SQLiteDatabase db = this.getWritableDatabase();
+        Cursor cursor = db.rawQuery("SELECT * FROM " + TABLE_NAME, null);
         return cursor;
+    }
+
+    public boolean deleteFood(int id) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        int result = db.delete(TABLE_NAME, COLUMN_ID + "=?", new String[]{String.valueOf(id)});
+        return result > 0;
     }
 }
